+++ rhue/src/io_schemes/mod.rs
@@ -1,0 +1,1 @@
+pub mod console;
