+++ rhue/src/main.rs
@@ -1,36 +1,19 @@
 use std::{
     env::args,
     fs::File,
-    io::{self, BufRead, Read},
+    io::{self, Read},
     result::Result,
 };
 
 mod ruleset;
 use ruleset::rule_engine::RuleEngine;
 use ruleset::{
-    rule::{Rule, RuleType},
-    rule_engine::RuleFuncs,
+    rule::{Rule, RuleType}
 };
 
-struct Console();
+mod io_schemes;
 
-impl RuleFuncs for Console {
-    fn print(output: &str) {
-        println!("{}", output);
-    }
-
-    fn input(prompt: &str) -> String {
-        if !prompt.is_empty() {
-            println!("{}", prompt);
-        }
-        let mut input = String::new();
-        io::stdin()
-            .lock()
-            .read_line(&mut input)
-            .expect("We should have had an input here");
-        input
-    }
-}
+use crate::io_schemes::console::ConsoleScheme;
 
 fn main() -> Result<(), io::Error> {
     let args: Vec<String> = args().skip(1).collect();
@@ -62,7 +45,7 @@ fn main() -> Result<(), io::Error> {
         })
         .collect::<Vec<Rule>>();
 
-    let mut rule_engine = RuleEngine::new(rules, program_file_text, Console());
+    let mut rule_engine = RuleEngine::new(rules, program_file_text, ConsoleScheme);
     let evaluated_text = rule_engine.evaluate();
 
     println!("{}", evaluated_text);
